Add project list
